preliminary work on datasources
